refactor: move OAuth2 TokenSource into api package with expiry-aware caching
Remove the standalone internal/oauth2 package, whose TokenSource cached a
token forever once fetched and only knew about a raw access token string.
Replace it with a TokenSource type in internal/api that talks to the
TokenManager directly via a TokenSourceClient interface, caches the
keyring.AccessToken's value and expiry, and re-fetches once the cached
token has expired (new isExpired helper with an injectable clock).
